integrate jar loader into test-suite.  works on mac in both file system and jar mode

+++ support-lib/java/com/dropbox/djinni/NativeLibLoader.java
@@ -19,26 +19,35 @@ package com.dropbox.djinni;
 import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
-import java.net.URI;
 import java.net.URISyntaxException;
+import java.net.URL;
+import java.nio.file.FileSystem;
+import java.nio.file.FileSystems;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.nio.file.StandardCopyOption;
+import java.util.Collections;
+import java.util.logging.Level;
+import java.util.logging.Logger;
+import java.util.stream.Stream;
 
 /**
  * Utilities for loading native libraries containing djinni interfaces
  * and records.  To load libraries in your application at startup,
  * simply place them:
- *  * inside the jar containing this code at the path `djinniNativeLibsJarPath`
+ *  * inside the jar containing this code at the (jar-relative)
+ *      path `djinniNativeLibsJarPath`
  *  * somewhere on the host filesystem and use the `djinniNativeLibsSysProp`
  *      system property to provide path(s)
- * 
- * To *disable* automatic loading of native libs at startup (and require
- * manual loading), use the `djinniLoadNativeLibsAtStartupProp` system
- * property.
+ *
+ * Then execute:
+ * <code>
+ *   NativeLibLoader.loadLibs()
+ * </code>
+ * to load the libraries (e.g. in your main() or a class static initializer). 
  */
-class NativeLibLoader {
+public class NativeLibLoader {
   
   /**
    * Canonical directory in a jar containing (djinni-adapted) native libraries
@@ -53,42 +62,20 @@ class NativeLibLoader {
   public static final String djinniNativeLibsSysProp =
       "djinni.native_libs_dirs";
   
-  /**
-   * System property flag indicating if native libraries should be loaded
-   * on program startup
-   */
-  public static final String djinniLoadNativeLibsAtStartupProp = 
-      "djinni.load_native_libs_at_startup";
+  private static final Logger log =
+      Logger.getLogger(NativeLibLoader.class.getName());
   
   private NativeLibLoader() { } 
-  
-  // Try to load libs at program startup
-  static {
-    String doLoad = System.getProperty(djinniLoadNativeLibsAtStartupProp);
-    if (doLoad == null || Boolean.valueOf(doLoad)) {
-      try {
-        loadLibs();
-      } catch (Exception e) {
-        throw new RuntimeException(e);
-      }
-    }
-  }
   
   // Load native libs from canonical locations
   public static void loadLibs() throws URISyntaxException, IOException {
     // Try to load from Jar
-    URI libs =
-        NativeLibLoader.class.getResource(djinniNativeLibsJarPath).toURI();
-    if (libs.getScheme().equals("jar")) {
-      loadLibFromJarPath(Paths.get(libs));
-    } else {
-      // Support running outside a jar
-      loadLibsFromLocalPath(Paths.get(libs));
-    }
+    loadLibsFromJarPath(djinniNativeLibsJarPath);
     
     // Try to load from system
     String localPaths = System.getProperty(djinniNativeLibsSysProp);
     if (localPaths != null) {
+      log.log(Level.FINE, "Loading local native libs");
       for (String localPath : localPaths.split(",")) {
         loadLibsFromLocalPath(Paths.get(localPath));
       }
@@ -100,12 +87,45 @@ class NativeLibLoader {
     File localFile = localPath.toFile();
     if (!localFile.exists()) { return; }
     if (localFile.isDirectory()) {
-      Files
-        .walk(localFile.toPath(), 1)
-        .forEach(p -> System.load(p.toFile().getAbsolutePath()));
+      log.log(Level.FINE, "Loading all libs in " + localFile.getAbsolutePath());
+      Stream<Path> streamPaths = Files.walk(localFile.toPath(), 1);
+      for (Path p : (Iterable<Path>)streamPaths::iterator) {
+        File f = p.toFile();
+        if (f.isFile()) {
+          loadLibrary(f.getAbsolutePath());
+        }
+      }
+      streamPaths.close();
     } else {
-      System.load(localFile.getAbsolutePath());
+      loadLibrary(localFile.getAbsolutePath());
     }
+  }
+  
+  // Load a directory of libs from a jar resource path
+  public static void loadLibsFromJarPath(String jarPath)
+      throws URISyntaxException, IOException {
+
+    // Do we have a valid path?
+    URL libsURL =
+        NativeLibLoader.class
+          .getClassLoader()
+          .getResource(djinniNativeLibsJarPath);
+    if (libsURL == null) { return; }
+    
+    // Are we actually referencing a jar path?
+    if (!libsURL.toURI().getScheme().equals("jar")) { return; }
+    
+    log.log(Level.FINE, "Loading libs from jar path " + jarPath);
+    
+    // Walk the directory and load libs
+    FileSystem fs =
+        FileSystems.newFileSystem(libsURL.toURI(), Collections.emptyMap());
+    Path myPath = fs.getPath(jarPath);
+    for (Path p : (Iterable<Path>)Files.walk(myPath, 1)::iterator) {
+      loadLibFromJarPath(p);
+    }
+    
+    fs.close();
   }
   
   // Load a single native lib from a jar resource with path `libPath`
@@ -117,12 +137,17 @@ class NativeLibLoader {
      */ 
     InputStream libIn =
         NativeLibLoader.class.getResourceAsStream(libPath.toString());
+    if (libIn == null) { return; }
     File tempLib =
         File.createTempFile(
             libPath.getName(libPath.getNameCount() - 1).toString(),
               // name tempfile after the lib to ease debugging
             null);
     tempLib.deleteOnExit();
+    
+    log.log(
+        Level.FINE,
+        "Copying jar lib " + libPath + " to " + tempLib.getAbsolutePath());
     try {
       Files.copy(libIn, tempLib.toPath(), StandardCopyOption.REPLACE_EXISTING);
     } catch (SecurityException e) { 
@@ -134,7 +159,11 @@ class NativeLibLoader {
             "path(s) using the system property " + djinniNativeLibsSysProp);
     }
     
-    System.load(tempLib.getAbsolutePath());
-  }  
+    loadLibrary(tempLib.getAbsolutePath());
+  }
+  
+  private static void loadLibrary(String abspath) {
+    System.load(abspath);
+    log.log(Level.INFO, "Loaded " + abspath);
+  }
 }
-
